% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.translate-package.R
\docType{package}
\name{aws.translate-package}
\alias{aws.translate-package}
\alias{aws.translate}
\title{aws.translate}
\description{
Client for AWS Translate
}
\details{
This is a client for AWS Translate, a machine translation service.
}
\seealso{
\href{http://aws.amazon.com/documentation/translate}{AWS Translate Documentation}

\code{\link{translate}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
