% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate text between languages}
\usage{
translate(text, from = "en", to, ...)
}
\arguments{
\item{text}{A character string containing a text}

\item{from}{A two-letter language code specifying the language of the source text.}

\item{to}{two-letter language code specifying the language to translate in to.}

\item{\dots}{Additional arguments passed to \code{\link{translateHTTP}}.}
}
\value{
A character string containing the translation.
}
\description{
Translate text between languages
}
\examples{
\dontrun{
  translate("hello world!", to = "fr")
}
}
